requirements.txt
google-adk
firebase-admin
google-api-core
google-cloud-storage
google-cloud-logging
opentelemetry-api
opentelemetry-exporter-gcp-trace
opentelemetry-sdk
pydantic
requests
configparser
uvicorn
fastapi
matplotlib
